Add Checkstyle and gradle support
