+++ src/main/java/duke/commands/Command.java
@@ -2,8 +2,8 @@ package duke.commands;
 
 import duke.info.exception.InvalidInputException;
 import duke.info.task.Calendar;
-import duke.ui.Ui;
 import duke.storage.Storage;
+import duke.ui.Ui;
 
 public abstract class Command {
