+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,12 +1,14 @@
 package duke.commands;
 
+import java.io.IOException;
+
 import duke.info.exception.InvalidInputException;
-import duke.ui.Ui;
 import duke.info.task.Calendar;
 import duke.storage.Storage;
+import duke.ui.Ui;
 import duke.utils.Text;
 
-import java.io.IOException;
+
 
 public class DeleteCommand extends Command {
 
@@ -39,7 +41,8 @@ public class DeleteCommand extends Command {
             ui.showTaskDeleted(taskString);
             storage.save(calendar);
         } catch (IndexOutOfBoundsException e) {
-            throw new InvalidInputException(String.format(Text.TEXT_INVALID_LIST_INDEX, calendar.numOfEntries(), Integer.parseInt(indexToDelete)));
+            throw new InvalidInputException(String.format(Text.TEXT_INVALID_LIST_INDEX,
+                    calendar.numOfEntries(), Integer.parseInt(indexToDelete)));
         } catch (NumberFormatException e) {
             throw new InvalidInputException(String.format(Text.TEXT_NON_INTEGER_LIST_INDEX, this.indexToDelete));
         } catch (IOException e) {
