+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,25 +1,25 @@
 package duke.commands;
 
 import duke.info.task.Calendar;
-import duke.ui.Ui;
 import duke.storage.Storage;
+import duke.ui.Ui;
 
-public class ListCommand extends Command {
+
+public class ExitCommand extends Command {
 
     /**
-     * Constructs a ListCommand
+     * Constructs an ExitCommand
      */
-    public ListCommand() {}
+    public ExitCommand() {}
 
     /**
-     * Displays the string representation of the specified calendar using the
-     * specified ui handler.
+     * Shows the goodbye message on the ui handler.
      * @param calendar - the calendar used in the program
      * @param ui - the ui handler for the program
      * @param storage - the storage handler for the program
      */
     @Override
     public void execute(Calendar calendar, Ui ui, Storage storage) {
-        ui.showCalendar(calendar);
+        ui.showGoodbye();
     }
 }
