+++ src/main/java/duke/commands/ListCommand.java
@@ -1,24 +1,25 @@
 package duke.commands;
 
 import duke.info.task.Calendar;
-import duke.ui.Ui;
 import duke.storage.Storage;
+import duke.ui.Ui;
 
-public class ExitCommand extends Command {
+public class ListCommand extends Command {
 
     /**
-     * Constructs an ExitCommand
+     * Constructs a ListCommand
      */
-    public ExitCommand() {}
+    public ListCommand() {}
 
     /**
-     * Shows the goodbye message on the ui handler.
+     * Displays the string representation of the specified calendar using the
+     * specified ui handler.
      * @param calendar - the calendar used in the program
      * @param ui - the ui handler for the program
      * @param storage - the storage handler for the program
      */
     @Override
     public void execute(Calendar calendar, Ui ui, Storage storage) {
-        ui.showGoodbye();
+        ui.showCalendar(calendar);
     }
 }
