+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,12 +1,14 @@
 package duke.commands;
 
+import java.io.IOException;
+
 import duke.info.exception.InvalidInputException;
 import duke.info.task.Calendar;
-import duke.ui.Ui;
 import duke.storage.Storage;
+import duke.ui.Ui;
 import duke.utils.Text;
 
-import java.io.IOException;
+
 
 public class MarkCommand extends Command {
 
@@ -39,7 +41,8 @@ public class MarkCommand extends Command {
             ui.showTaskComplete(calendar.taskStringAtIndex(index));
             storage.save(calendar);
         } catch (IndexOutOfBoundsException e) {
-            throw new InvalidInputException(String.format(Text.TEXT_INVALID_LIST_INDEX, calendar.numOfEntries(), Integer.parseInt(indexToMark)));
+            throw new InvalidInputException(String.format(Text.TEXT_INVALID_LIST_INDEX,
+                    calendar.numOfEntries(), Integer.parseInt(indexToMark)));
         } catch (NumberFormatException e) {
             throw new InvalidInputException(String.format(Text.TEXT_NON_INTEGER_LIST_INDEX, this.indexToMark));
         } catch (IOException e) {
