+++ src/main/java/duke/commands/UnmarkCommand.java
@@ -1,12 +1,12 @@
 package duke.commands;
 
+import java.io.IOException;
+
 import duke.info.exception.InvalidInputException;
-import duke.ui.Ui;
 import duke.info.task.Calendar;
 import duke.storage.Storage;
+import duke.ui.Ui;
 import duke.utils.Text;
-
-import java.io.IOException;
 
 public class UnmarkCommand extends Command {
 
@@ -39,7 +39,8 @@ public class UnmarkCommand extends Command {
             ui.showTaskIncomplete(calendar.taskStringAtIndex(index));
             storage.save(calendar);
         } catch (IndexOutOfBoundsException e) {
-            throw new InvalidInputException(String.format(Text.TEXT_INVALID_LIST_INDEX, calendar.numOfEntries(), Integer.parseInt(indexToUnmark)));
+            throw new InvalidInputException(String.format(Text.TEXT_INVALID_LIST_INDEX,
+                    calendar.numOfEntries(), Integer.parseInt(indexToUnmark)));
         } catch (NumberFormatException e) {
             throw new InvalidInputException(String.format(Text.TEXT_NON_INTEGER_LIST_INDEX, this.indexToUnmark));
         } catch (IOException e) {
