+++ src/main/java/duke/info/task/Task.java
@@ -1,8 +1,8 @@
 package duke.info.task;
 
 import java.time.LocalDate;
-import java.time.format.DateTimeParseException;
 import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
 
 public abstract class Task {
