+++ src/main/java/duke/utils/Text.java
@@ -8,17 +8,21 @@ public class Text {
 
     public static final String TEXT_LOADING_ERROR = "No previous calendar found... Creating new save file";
 
-    public static final String TEXT_INVALID_COMMAND_ERROR = "Sorry! I don't know what that means :( Please type a valid command!";
+    public static final String TEXT_INVALID_COMMAND_ERROR = "Sorry! I don't know what that means :("
+            + "Please type a valid command!";
 
     public static final String TEXT_GOODBYE_MESSAGE = "Bye! Hope to see you again soon... :)";
 
     public static final String TEXT_EMPTY_LIST = "Sorry! There are no items in the list as of now! :(";
 
-    public static final String TEXT_INVALID_LIST_INDEX = "Sorry! This list only has %s values but the input index for this command was %s";
+    public static final String TEXT_INVALID_LIST_INDEX = "Sorry! This list only has %s values but"
+            + " the input index for this command was %s";
 
-    public static final String TEXT_NON_INTEGER_LIST_INDEX = "Sorry! The list index should be an integer value but the input value for this command was %s";
+    public static final String TEXT_NON_INTEGER_LIST_INDEX = "Sorry! The list index should be an integer value but"
+            + " the input value for this command was %s";
 
-    public static final String TEXT_TASK_ADDED = "Got it! I've added this task:\n   %s\nNow you have %s tasks in the list.";
+    public static final String TEXT_TASK_ADDED = "Got it! I've added this task:\n   %s\n"
+            + "Now you have %s tasks in the list.";
 
     public static final String TEXT_MARKED = "Nice! I've marked this task as complete:\n   %s";
